LGTM

+++ programs/defios/src/helper.rs
@@ -1,7 +1,7 @@
-pub fn calculate_mint(buy: u128, token_supply: u128) -> u64 {
+pub fn calculate_mint(buy: u64, token_supply: u64) -> u64 {
     return 1;
 }
 
-pub fn calculate_burn(sell: u128, token_supply: u128) -> u64 {
+pub fn calculate_burn(sell: u64, token_supply: u64) -> u64 {
     return 1;
 }
